pyproject.toml
[project]
name = "zero-dce-keras"
version = "0.1.0"
description = "Zero-DCE: Zero-Reference Deep Curve Estimation for low-light image enhancement in Keras 3"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "keras>=3.11.3",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "pyqt6>=6.10.0",
    "pytest>=8.4.2",
    "pytest-qt>=4.4.0",
    "retinex>=0.0.1",
    "tensorflow[and-cuda]>=2.20.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.0",
    "ruff>=0.14.6",
]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by ruff-format)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"test_*.py" = ["F401", "F811"]

[tool.ruff.lint.isort]
known-third-party = ["keras", "tensorflow", "numpy", "cv2", "PIL", "matplotlib", "pytest", "PyQt6"]
